+++ tests/run-tests.js
@@ -1,2 +1,3 @@
 import './scoringEngine.test.js';
 import './sessionSetup.test.js';
+import './sessionActions.test.js';